customer-ms: modified servivcies

+++ Planealo/negocio/Customers/src/main/java/com/planealo/negocio/customers/service/impl/CustomerServiceImpl.java
@@ -11,7 +11,6 @@ import com.planealo.negocio.customers.models.entity.Customer;
 import com.planealo.negocio.customers.repository.CustomerRepository;
 import com.planealo.negocio.customers.service.ICustomerService;
 
-
 @Service
 public class CustomerServiceImpl implements ICustomerService<Customer, String> {
 
@@ -28,11 +27,13 @@ public class CustomerServiceImpl implements ICustomerService<Customer, String> {
 	}
 
 	@Override
-	public boolean delete(String ref) {
-		Optional<Customer> customer = this.customerRepo.findByReferencia(ref);
+	public Boolean delete(Customer customerDelete) {
+		Optional<Customer> customer = this.customerRepo.findByReferencia(customerDelete.getReferencia());
 		if (customer.isPresent()) {
-			this.customerRepo.delete(customer.get());
-			return true;
+			if (customer.get().getPassword().equals(customerDelete.getPassword())) { // Puede que se tenga que desencriptar
+				this.customerRepo.delete(customer.get());
+				return true;
+			}
 		}
 		return false;
 	}
@@ -49,27 +50,28 @@ public class CustomerServiceImpl implements ICustomerService<Customer, String> {
 	}
 
 	@Override
-	public boolean addAll(List<Customer> t) {
+	public Boolean addAll(List<Customer> t) {
 		this.customerRepo.saveAll(t);
 		return false;
 	}
 
 	@Override
-	// @Transactional -> https://danielme.com/2023/02/22/curso-spring-data-jpa-transacciones-propagacion-excepciones/
-	public Customer editUser(String ref, Customer t) {
-		 Optional<Customer> existingCustomer = this.customerRepo.findByReferencia(ref);
-		 if(existingCustomer.isPresent()) {
-			 // Verifica que el correo no exista
-	        if (this.customerRepo.findByEmail(t.getEmail()).isEmpty()) {
-	            Customer customer = existingCustomer.get();
-	            customer.setNombre(t.getNombre());
-	            customer.setEmail(t.getEmail());
-	            customer.setPassword(t.getPassword());
-	            customer.setFechaUltimaActualizacion(LocalDateTime.now());
-	            return this.customerRepo.save(customer);
-	        }
-		 }
-        return null;
+	// @Transactional ->
+	// https://danielme.com/2023/02/22/curso-spring-data-jpa-transacciones-propagacion-excepciones/
+	public Customer editUser( Customer t) {
+		Optional<Customer> existingCustomer = this.customerRepo.findByReferencia(t.getReferencia());
+		if (existingCustomer.isPresent()) {
+			// Verifica que el correo no exista
+			if (this.customerRepo.findByEmail(t.getEmail()).isEmpty()) {
+				Customer customer = existingCustomer.get();
+				customer.setNombre(t.getNombre());
+				customer.setEmail(t.getEmail());
+				customer.setPassword(t.getPassword());
+				customer.setFechaUltimaActualizacion(LocalDateTime.now());
+				return this.customerRepo.save(customer);
+			}
+		}
+		return null;
 	}
 
 	@Override
